Updated build targets.
* Added win11-arm64 support.
* Dropped python 3.10 support and added python 3.14 support.
* Bumped packaging versions.

+++ example/jls/read_fuzzer.c
@@ -167,7 +167,7 @@ int on_read_fuzzer(struct app_s * self, int argc, char * argv[]) {
             }
             if (*(guard - 1) == guard_byte) {
                 printf("incomplete: ");
-                for (int32_t i = -guard_length; i < 0; ++i) {
+                for (int32_t i = - ((int32_t) guard_length); i < 0; ++i) {
                     printf(" %02x", guard[i]);
                 }
                 break;
